Avance y fix de problemas en modulo de venta

+++ protected/controllers/VentaController.php
@@ -6,7 +6,7 @@ class VentaController extends Controller
 	 * @var string the default layout for the views. Defaults to '//layouts/column2', meaning
 	 * using two-column layout. See 'protected/views/layouts/column2.php'.
 	 */
-	 public $layout='//layouts/intraLayout';
+	public $layout='//layouts/intraLayout';
 
 	/**
 	 * @return array action filters
@@ -32,7 +32,7 @@ class VentaController extends Controller
 				'users'=>array('*'),
 			),
 			array('allow', // allow authenticated user to perform 'create' and 'update' actions
-				'actions'=>array('index','view','create','update'),
+				'actions'=>array('create','update','index','view', 'select', 'select2', 'eliminar'),
 				'users'=>array('@'),
 			),
 			array('allow', // allow admin user to perform 'admin' and 'delete' actions
@@ -51,11 +51,42 @@ class VentaController extends Controller
 	 */
 	public function actionView($id)
 	{
+		$model=$this->loadModel($id);
+		$model3=new Propiedad;
+		$model3=Propiedad::model()->findByPk($model->id_propiedad);
 		$this->render('view',array(
 			'model'=>$this->loadModel($id),
+			'model3'=>$model3
 		));
 	}
 
+	public function actionSelect()
+	{
+		$model=new Venta('search');
+		$model->unsetAttributes();  // clear any default values
+		if(isset($_GET['Venta']))
+			$model->attributes=$_GET['Venta'];
+		$this->render('select',array('model'=>$model));
+	}
+
+	public function actionSelect2()
+	{
+		$model=new Venta('search');
+		$model->unsetAttributes();  // clear any default values
+		if(isset($_GET['Venta']))
+			$model->attributes=$_GET['Venta'];
+		$this->render('select2',array('model'=>$model));
+	}
+
+	public function actionEliminar($id)
+	{
+		$model=$this->loadModel($id);
+		$model3=new Propiedad;
+		$model3=Propiedad::model()->findByPk($model->id_propiedad);
+		$this->render('eliminar',array(
+			'model'=>$model, 'model3'=>$model3
+		));
+	}
 	/**
 	 * Creates a new model.
 	 * If creation is successful, the browser will be redirected to the 'view' page.
@@ -90,7 +121,6 @@ class VentaController extends Controller
 				'defaultOrder'=>array('id_propiedad'=>true),
 			),
 		));
-
 		// Uncomment the following line if AJAX validation is needed
 		// $this->performAjaxValidation($model);
 
@@ -98,14 +128,21 @@ class VentaController extends Controller
 		{
 			$model->attributes=$_POST['Venta'];
 			$model->rut_admin = Yii::app()->session['admin_rut'];
-			$model3=Propiedad::model()->findByPk($model->id_propiedad);
-			if($model3->activo_propiedad == 1){
-				$model3->activo_propiedad= 0;
-				if($model->save() && $model3->save())
-				{
-					$this->redirect(array('view','id'=>$model->id_venta));
+			if ($model->id_propiedad != '') {
+				$model3=Propiedad::model()->findByPk($model->id_propiedad);
+				if($model3->activo_propiedad == 1){
+					$model3->activo_propiedad= 0;
+					$valor = intval(preg_replace('/[^0-9]+/', '', $model->ganancia_venta),10);
+					$model->ganancia_venta = $valor;
+					if($model->save() && $model3->save())
+					{
+						Yii::app()->user->setFlash('success','La venta fue registrada exitosamente.');
+						$this->redirect(array('view','id'=>$model->id_venta));
+					}
+				}else {
+					Yii::app()->user->setFlash('error','La propiedad ya se encuentra con un servicio prestado.');
 				}
-			}else {
+			}else{
 				Yii::app()->user->setFlash('error','La propiedad ya se encuentra con un servicio prestado.');
 			}
 		}
@@ -126,19 +163,67 @@ class VentaController extends Controller
 	public function actionUpdate($id)
 	{
 		$model=$this->loadModel($id);
+		$variable= $this->loadModel($id);
+		$model3=new Propiedad;
+		$model3=Propiedad::model()->findByPk($model->id_propiedad);
+
 
 		// Uncomment the following line if AJAX validation is needed
 		// $this->performAjaxValidation($model);
 
 		if(isset($_POST['Venta']))
 		{
-			$model->attributes=$_POST['Venta'];
-			if($model->save())
+			$model->attributes=$_POST['Arriendo'];
+			$model3=Propiedad::model()->findByPk($model->id_propiedad);
+			if($model->id_propiedad != $variable->id_propiedad){
+				$model4=new Propiedad;
+				$model4=Propiedad::model()->findByPk($variable->id_propiedad);
+				$model4->activo_propiedad=1;
+				if($model4->save()){
+					Yii::app()->user->setFlash('success','La propiedad enlazada con el arriendo fue modificada.');
+				}else {
+					Yii::app()->user->setFlash('danger','La propiedad enlazada con el arriendo no pudo ser modificada.');
+				}
+			}
+			$model3->activo_propiedad =0;
+			if($model->save() && $model3->save()){
+				Yii::app()->user->setFlash('success','El venta fue actualizado.');
 				$this->redirect(array('view','id'=>$model->id_venta));
+			}else{
+				Yii::app()->user->setFlash('danger','El venta no fue actualizado.');
+			}
 		}
+		$criteria2 = new CDbCriteria();
+		$criteria2->condition='activo_arrendatario=1';
+
+		$dataProvider=new CActiveDataProvider(Arrendatario::model(), array(
+			'keyAttribute'=>'rut_arrendatario',// IMPORTANTE, para que el CGridView conozca la seleccion
+			'criteria'=>$criteria2,
+			'pagination'=>array(
+				'pageSize'=>5,
+			),
+			'sort'=>array(
+				'defaultOrder'=>array('rut_arrendatario'=>true),
+			),
+		));
+		$criteria = new CDbCriteria();
+		$criteria->condition='activo_propiedad=1 AND eliminado_propiedad=0';
+		$dataProvider2=new CActiveDataProvider(Propiedad::model(), array(
+			'keyAttribute'=>'id_propiedad',// IMPORTANTE, para que el CGridView conozca la seleccion
+			'criteria'=>$criteria,
+			'pagination'=>array(
+				'pageSize'=>5,
+			),
+			'sort'=>array(
+				'defaultOrder'=>array('id_propiedad'=>true),
+			),
+		));
 
 		$this->render('update',array(
 			'model'=>$model,
+			'model3'=>$model3,
+			'dataProvider'=>$dataProvider,
+			'dataProvider2'=>$dataProvider2,
 		));
 	}
 
@@ -149,8 +234,12 @@ class VentaController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
-
+		$model =$this->loadModel($id);
+		$model2 = Propiedad::model()->findByPk($model->id_propiedad);
+		$model2->activo_propiedad = 1;
+		if($model2->save()){
+			$this->loadModel($id)->delete();
+		}
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
 		if(!isset($_GET['ajax']))
 			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
